Tests: Fix errors in remaining unit tests. Use semantic methods instead of fixed-width methods, use defined type widths instead of magic type widths, misc cleanup.

+++ src/stateserver/DistributedObject.cpp
@@ -86,8 +86,8 @@ void DistributedObject::append_required_data(Datagram &dg, bool client_only, boo
 	dg.add_doid(m_do_id);
 	dg.add_location(m_parent_id, m_zone_id);
 	dg.add_uint16(m_dclass->get_number());
-	uint32_t field_count = m_dclass->get_num_inherited_fields();
-	for(uint32_t i = 0; i < field_count; ++i)
+	size_t field_count = m_dclass->get_num_inherited_fields();
+	for(size_t i = 0; i < field_count; ++i)
 	{
 		DCField *field = m_dclass->get_inherited_field(i);
 		if(field->is_required() && !field->as_molecular_field() && (!client_only
